Remove vim swp file and fix cloud-init
This patch removes an accidentally checked-in vim .swp file and fixes
cloud-init.  The definition for this CentOS70 (C7) image was copied from the
CentOS65 (C65) image.  In the cleanup script, the package "freetype" was being
removed.  In C65, this worked fine as there were no dependancy issues.
However, in C7, the cloud-init package has a dependancy on freetype,
thus removing it removes cloud-init, which has adverse affects.