Remove nfs-common install
This patch removes nfs-common install from the veewee build processes
for the Ubuntu-12.04-Precise image.  This package is causing the veewee
build to fail (likely due to mirrors being messed up).  However, this
really is an optional package, so it is reasonable to leave it to the
user to install this package, should the actually want it.