Clear yum fastest mirror cache
Yum caches which mirror it deems is fastest.  This patch removes this
cache from yum based distros so that the first time the user runs yum,
it will re-determine which mirror is the fastest, which could easily
have changed from image creation time to the time a VM is spun up from
this image.