Remove resize2fs hack
Debian 7 (Wheezy) had a version of resize2fs (e2fsprogs package) that
was poorly optimized and caused long resizes.  To address this, a static
compiled version of resize2fs with the optimization fix was added.
However, in Debian 8 (Jessie), this package has been updated and has a
version of resize2fs that contains the optimization fix.  As such, we no
longer need this static file.