Ensure EPEL is working for Centos7
This patch ensures that the EPEL repo is working on the Centos 7.0
image.  There were conflicting lines on dhc.sh and cleanup.sh.  dhc.sh
was installing epel (as was base.sh...) and cleanup.sh was removing it.
This patch assumes we do indeed want EPEL enabled and removes the
deletion of the repo from the cleanup.sh script.