Ubuntu-14.04: close compiled preseed file
After processsing the template, the preseed.cfg file is supposed to be
written to disk, then served via veewee.  However, it appears as though
ruby 2.1.0 is leaving the file handle open and not flushing to disk, so
the subsequent serving of preseed.cfg just serves a blank/empty file.
This patch adds an out.close call to ensure that the file is flushed to
disk.