All: Remove update_hostname/etc_hosts modules
This patch removes the update_hostname and update_etc_hosts cloud-init
modules in order to allow for the user to provide a hostname and it not
get wiped out by cloud-init on the subsequent boot(s).  Also included is
a cleanup snippet that removes the hostname used during image buildtime
from the /etc/hosts file.